refactor base actor
