+++ tests/test_single_controller.py
@@ -4,12 +4,9 @@
 import logging
 import os
 import pathlib
-from datetime import timedelta
-from typing import Optional
 
 import pytest
 import ray
-import torch
 import torch.distributed as dist
 from transformers import (
     AutoModelForCausalLM,
@@ -19,93 +16,17 @@ from transformers import (
 
 from compose_rl.algorithms.online.generation_utils import (
     create_vllm_engines,
-    init_process_group,
 )
-from compose_rl.utils.ray_utils import (
-    get_free_port,
-    get_node_ip,
-    is_cuda_visible_devices_set,
-    start_ray_server,
-)
-from tests.common import world_size
+from compose_rl.utils.ray_utils import start_ray_server
+from tests.common import world_size, BaseDistributedGPUActor
 
 # Set up logging
 logger = logging.getLogger(__name__)
 
 
 @ray.remote(num_gpus=1)
-class DistributedGPUActor:
-
-    def __init__(
-        self,
-        rank: int,
-        world_size: int,
-        master_addr: Optional[str] = None,
-        master_port: Optional[int] = None,
-    ):
-        """Initialize the distributed GPU actor.
-
-        Args:
-            rank: The rank of this process in the distributed group
-            world_size: Total number of processes in the distributed group
-            master_addr: Master node address. If None, will allocate dynamically for rank 0
-            master_port: Master node port. If None, will allocate dynamically for rank 0
-        """
-        self.rank = rank
-        self.world_size = world_size
-        self.master_addr = master_addr
-        self.master_port = master_port
-
-        # Set up basic environment variables
-        os.environ['WORLD_SIZE'] = str(world_size)
-        os.environ['RANK'] = str(rank)
-
-        # Set LOCAL_RANK based on Ray GPU allocation
-        os.environ['LOCAL_RANK'] = '0' if is_cuda_visible_devices_set(
-        ) else str(ray.get_gpu_ids()[0])
-
-        # If this is rank 0 and no master_addr/master_port provided, allocate them
-        if rank == 0 and (master_addr is None or master_port is None):
-            self._allocate_master_address()
-
-        os.environ['MASTER_ADDR'] = self.master_addr  # type: ignore
-        os.environ['MASTER_PORT'] = str(self.master_port)  # type: ignore
-
-        self.model = None
-        self.model_update_group = None
-
-    def _allocate_master_address(self):
-        """Allocate master address and port for rank 0."""
-        if self.master_addr is None:
-            # Get the local IP address
-            self.master_addr = get_node_ip()
-
-        if self.master_port is None:
-            # Allocate a free port
-            self.master_port = get_free_port()
-
-    def get_master_address(self) -> tuple[Optional[str], Optional[int]]:
-        """Return the master address and port as a tuple."""
-        return (self.master_addr, self.master_port)
-
-    def get_free_port(self):
-        return get_free_port()
-
-    def init_train_process_group(self):
-        """Initialize the distributed process group."""
-        # Initialize process group
-        dist.init_process_group(timeout=timedelta(seconds=30))
-        logger.info(f'is distributed initialized: {dist.is_initialized()}')
-        # Print debug information
-        num_visible_devices = torch.cuda.device_count()
-        logger.info(f'num_visible_devices: {num_visible_devices}')
-        logger.info('Ray actor init envs:')
-        logger.info(f'rank: {dist.get_rank()}')
-        logger.info(f'node_rank: {dist.get_rank() // 8}')
-        logger.info(f'world_size: {dist.get_world_size()}')
-        logger.info(f'local_rank: {dist.get_rank() % 8}')
-        logger.info(f'master_addr: {self.master_addr}')
-        logger.info(f'master_port: {self.master_port}')
+class DistributedGPUActor(BaseDistributedGPUActor):
+    """Distributed GPU actor for testing."""
 
     def init_model(self, model_name: str):
         """Initialize the model."""
@@ -131,36 +52,6 @@ class DistributedGPUActor:
             # after the remote calls are initialized above
             dist.broadcast(p, src=0, group=self.model_update_group)
             ray.get(refs)
-
-    def tensor_all_reduce(self) -> float:
-        """Perform a simple tensor all_reduce operation."""
-        # Create a tensor on the GPU and perform all_reduce
-        device = torch.device('cuda')
-        x = torch.ones(1, device=device, dtype=torch.int32)
-        dist.all_reduce(x)
-
-        return x.item()
-
-    def init_vllm_process_group(
-        self,
-        backend: str,
-        master_addr: str,
-        master_port: int,
-        world_size: int,
-        rank: int,
-        group_name: str,
-    ):
-        """Initialize the process group on trainer rank 0 and vllm engines."""
-        # NOTE vLLM seems to have a safer implementation of init_process_group:
-        # https://github.com/vllm-project/vllm/blob/v0.9.1/examples/offline_inference/rlhf.py#L105
-        # we should look into using that instead
-        self.model_update_group = init_process_group(
-            backend=backend,
-            init_method=f'tcp://{master_addr}:{master_port}',
-            world_size=world_size,
-            rank=rank,
-            group_name=group_name,
-        )
 
 
 @pytest.mark.gpu
@@ -228,7 +119,7 @@ def test_distributed_ray_actors(
 
             # Perform tensor all_reduce on all actors
             reduce_tasks = [
-                actor.tensor_all_reduce.remote()  # type: ignore
+                actor.test_tensor_all_reduce.remote()  # type: ignore
                 for actor in train_actors
             ]
             results = ray.get(reduce_tasks)
@@ -275,7 +166,7 @@ def test_distributed_ray_actors(
                 ) for i, engine in enumerate(vllm_engines)
             ]
             refs.append(
-                master_actor.init_vllm_process_group.remote(  # type: ignore
+                master_actor.add_process_group.remote(  # type: ignore
                     backend='nccl',
                     master_addr=master_addr,
                     master_port=new_port,
